+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 0.3.0
+- Support PublishAsync (experimental feature)
+- Simply return boolean on Publish function to indicate if there is any subscribed handlers to the published event.
+
 ## 0.2.2
 - Fixed event unsubscription not working with event wrapper.
 - Add basic unit tests for the event bus.